SQL-2293: Create separate direct cluster (DC) integration test file
